Refactors layout and adds new sidebar component
This commit introduces a new sidebar component for improved navigation and a cleaner layout structure.

The General layout and related code have been refactored to utilize this new component.  Styling updates have also been applied to enhance the visual appeal of the application, including a new login illustration and color palette.  The main App component is simplified and now uses the new GeneralLayout for routing.  The router has been updated to include routes for a user module, with a list view, and a dashboard route.  Additionally, the project now uses Sass for styling.